2903 - update NR to use legal api proxy endpoint
